# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/trips": {
      "get": {
        "operationId": "TripsController_findAll",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "List all trips",
        "tags": [
          "Trips"
        ]
      }
    },
    "/api/trips/{id}": {
      "get": {
        "operationId": "TripsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Get trip by ID",
        "tags": [
          "Trips"
        ]
      }
    },
    "/api/trips/{id}/reassign": {
      "patch": {
        "operationId": "TripsController_reassign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "driver": {
                    "type": "string",
                    "example": "Raghav Reddy"
                  }
                },
                "required": [
                  "driver"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Reassign driver for a trip",
        "tags": [
          "Trips"
        ]
      }
    },
    "/api/trips/{id}/status": {
      "patch": {
        "description": "Updates trip status and syncs the linked delivery request status so other actor portals show the same state.",
        "operationId": "TripsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "example": "Accepted",
                    "enum": [
                      "Assigned",
                      "Accepted",
                      "Rejected",
                      "In Transit",
                      "Picked Up",
                      "Delivered",
                      "Completed",
                      "Delayed",
                      "Queued",
                      "Issue Reported"
                    ]
                  }
                },
                "required": [
                  "status"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Update trip status",
        "tags": [
          "Trips"
        ]
      }
    },
    "/api/trips/{id}/report-issue": {
      "patch": {
        "description": "Reports an issue for a trip, updates the linked delivery request, and creates notifications for other portals.",
        "operationId": "TripsController_reportIssue",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-user-role",
            "in": "header",
            "required": true,
            "description": "Role of the requesting user (superuser | fleet-manager | business-client | driver)",
            "schema": {
              "type": "string",
              "enum": [
                "superuser",
                "fleet-manager",
                "business-client",
                "driver"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "required": false,
            "description": "ID of the requesting user (e.g. SU-001). Used for self-service endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "issueType": {
                    "type": "string",
                    "example": "Vehicle problem"
                  },
                  "description": {
                    "type": "string",
                    "example": "Vehicle broke down near pickup location."
                  },
                  "status": {
                    "type": "string",
                    "example": "Issue Reported"
                  }
                },
                "required": [
                  "issueType",
                  "description"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Report an issue for a trip",
        "tags": [
          "Trips"
        ]
      }
    },
...
  },
  "info": {
    "title": "DeliverSync API",
    "description": "REST API for the DeliverSync delivery management platform. Role-based access is controlled via the `x-user-role` request header. Supported roles: superuser, fleet-manager, business-client, driver.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}